Move helpers to testing dir.
